users

+++ src/controllers/userController.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { userModel } from "../models/userModel";
+import { AuthRequest } from "../middleware/authMiddleware";
 
 class UserController{
 
@@ -36,7 +37,7 @@ class UserController{
             }
     }
 
-    async update(req: Request, res: Response) {
+    async update(req: AuthRequest, res: Response) {
         try {
             const { username, email, password } = req.body;
             const userId = req.params.userId;
@@ -44,8 +45,14 @@ class UserController{
                 res.status(400).send(`userId, username, email or password are required (full replace)`);
                 return;
             }
-            else if (!await userModel.findById(userId)) {
+            const user = await userModel.findById(userId);
+            if (!user) {
                 res.status(404).send(`User not found`);
+                return;
+            }
+            if(req.user && req.user._id !== userId){
+                res.statusCode = 403;
+                res.status(403).send(`Forbidden - You can update only your own user data`);
                 return;
             }
             const userExists = await userModel.findOne({ $or: [ { username }, { email } ] });
@@ -54,26 +61,36 @@ class UserController{
                 return;
             }
 
-            const user = await userModel.findByIdAndUpdate(userId, { userId, username, email, password }, { new: true });
+            const updatedUser = await userModel.findByIdAndUpdate(userId, { userId, username, email, password }, { new: true });
             if (!user) {
                 res.status(404).send(`User not found`);
                 return;
             }
-            res.json(user);
+            res.json(updatedUser);
         } catch (err: any) {
             res.status(500).send(err.message);
         }
     }
 
-    async delete(req: Request, res: Response) {
+    async delete(req: AuthRequest, res: Response) {
         try {
             const { userId } = req.params;
-            const user = await userModel.findByIdAndDelete(userId);
+            const user = await userModel.findById(userId);
             if (!user) {
                 res.status(404).send("Not Found - User does not exist");
                 return;
             }
-            res.status(200).send(`User with id ${userId} deleted successfully`);
+            if(req.user && req.user._id !== userId){
+                res.statusCode = 403;
+                res.status(403).send(`Forbidden - You can delete only your own user`);
+                return;
+            }
+            const deletedUser = await userModel.findByIdAndDelete(userId);
+            if (!deletedUser) {
+                res.status(404).send("Not Found - User does not exist");
+                return;
+            }
+            res.status(200).send(`User with id ${userId} deleted successfully: ${JSON.stringify(deletedUser)}`);
         } catch (err: any) {
             res.status(500).send(err.message);
         }
